updated MachineDAO

+++ pujcovnaStroju/src/main/java/cz/muni/fi/pa165/pujcovnastrojuDAO/MachineDAOImpl.java
@@ -3,8 +3,6 @@ package cz.muni.fi.pa165.pujcovnastrojuDAO;
 import java.util.List;
 
 import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Persistence;
 import javax.persistence.TypedQuery;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
@@ -18,21 +16,34 @@ public class MachineDAOImpl implements MachineDAO{
 	
 	private EntityManager entityManager;
 	
-	public MachineDAOImpl() {
-		 EntityManagerFactory emf = Persistence.createEntityManagerFactory("TestPU");
-	     entityManager = emf.createEntityManager();
+	public MachineDAOImpl(EntityManager em) {
+	     if (em == null) {
+	    	 throw new IllegalArgumentException("arbument 'em' must be set");
+	     }
+	     entityManager = em;
 	}
 	
+	public MachineDAOImpl() {}
+	
 	public Machine create(Machine machine) {
-		entityManager.getTransaction().begin();
-		//TODO check values
+		if (machine == null) {
+			throw new IllegalArgumentException("unset argument 'machine");
+		}
+		if (machine.getLabel() == null) {
+			throw new IllegalArgumentException("argument 'label' must be set");
+		}
+		
 		entityManager.persist(machine);
-		entityManager.getTransaction().commit();
 		return machine;
 	}
 
 	public Machine update(Machine machine) {
-		entityManager.getTransaction().begin();
+		if (machine == null) {
+			throw new IllegalArgumentException("unset argument 'machine'");
+		}
+		if (machine.getId() == null) {
+			throw new IllegalArgumentException("argument 'id' must be set");
+		}
 		Machine toBeUpdated = entityManager.find(Machine.class, machine.getId());
 		if (toBeUpdated != null) {
 			toBeUpdated.setLabel(machine.getLabel());
@@ -42,42 +53,41 @@ public class MachineDAOImpl implements MachineDAO{
 			toBeUpdated = machine;
 			entityManager.persist(machine);
 		}
-		entityManager.getTransaction().commit();
 		return toBeUpdated;
 	}
 
 	public Machine read(Long id) {
-		entityManager.getTransaction().begin();
+		if (id == null) {
+			throw new IllegalArgumentException("unset argument 'id'");
+		}
 		Machine result = entityManager.find(Machine.class, id);
-		entityManager.getTransaction().commit();
 		return result;
 	}
 
 	public void delete(Machine machine) {
-		entityManager.getTransaction().begin();
+		if (machine == null) {
+			throw new IllegalArgumentException("unset argument 'machine'");
+		}
 		Machine toBeDeleted = entityManager.find(Machine.class, machine.getId());
 		if (toBeDeleted != null) {
 			entityManager.remove(toBeDeleted);
 		}
-		entityManager.getTransaction().commit();
 	}
 
-	public List<Machine> getAllMachines() {
-		entityManager.getTransaction().begin();
-		
+	public List<Machine> getAllMachines() {		
 		CriteriaBuilder cb = entityManager.getCriteriaBuilder();
 		CriteriaQuery<Machine> cq = cb.createQuery(Machine.class);
 		Root<Machine> machines = cq.from(Machine.class);
 		cq.select(machines);
 		TypedQuery<Machine> q = entityManager.createQuery(cq);
 		
-		entityManager.getTransaction().commit();
 		return q.getResultList();
 	}
 
 	public List<Machine> getMachinesByType(MachineTypeEnum type) {
-		entityManager.getTransaction().begin();
-		
+		if (type == null) {
+			throw new IllegalArgumentException("unset argument 'type");
+		}
 		CriteriaBuilder criteriaBuilder = entityManager.getCriteriaBuilder();
 		CriteriaQuery<Machine> criteriaQuery = criteriaBuilder.createQuery(Machine.class);
 		Root<Machine> rootMachines = criteriaQuery.from(Machine.class);
@@ -89,7 +99,6 @@ public class MachineDAOImpl implements MachineDAO{
 		q.setParameter(param, type);
 		List<Machine> machineList= q.getResultList();
 		
-		entityManager.getTransaction().commit();
 		return machineList;
 	}
